requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0  # Added for Gemini
chromadb>=0.4.18
sentence-transformers>=2.2.2
pytube>=15.0.0
openai-whisper>=20231117
pydub>=0.25.1
transformers>=4.35.0
torch>=2.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
nltk>=3.8.0
spacy>=3.7.0
streamlit-chat>=0.1.0
tiktoken>=0.5.0
langchain>=0.1.0
langchain-text-splitters>=0.0.1
python-multipart>=0.0.6
hashlib